==> JuliaAcademy/DataFrames/5_collecting_experiments_data_in_a_data_frame.ipynb <==
# Collecting experiments data in a data frame

using DataFrames
using Statistics
using PyPlot
using Random
using Pipe
using Test

# U(0,1)
# How many draws till the sum exceeds 1?
function sim_e()
    draw = Float64[]
    while true
        push!(draw, rand())
        sum(draw) > 1.0 && return draw
    end
end

Random.seed!(1234) # want to have the same result

draws = [sim_e() for _ in 1:5]

# Check the sum of every draw is greater than 1.0
@test all(@. sum(draws) > 1.0)

# Check it is the last element exceeds 1.0
@test all(@. (sum(draws) - last(draws)) < 1.0)

# Put lots of experiments into a data frame
df = DataFrame()

@time for i in 1:10^6
    push!(df, (id = i, pos = sim_e()))
end

df

# This form produces a :jumps column with value 1,000,000 in every row
# transform(df, :pos => length => :jumps)

# This form tells transform to go row-by-row.
transform!(df, :pos => ByRow(length) => :jumps)

mean(df.jumps)

# or 
combine(df, :jumps => mean => :jumps_mean)

# close to e
MathConstants.e

# Find a distribution of number of jumps
jumps_agg = @pipe df |> groupby(_, :jumps, sort=true) |> combine(_, :jumps => length)

# 1. normalize by sum
# 2. calculate theoretic values
transform(
    jumps_agg,     
    :jumps_length => (x -> x ./ sum(x))             => :simulation,
    :jumps        => ByRow(x -> (x-1)/factorial(x)) => :theory
)

df

df_test = select(
    df,
    :pos => ByRow(first) => :first,
    :pos => ByRow(last)  => :last
)

hist(df_test.first, 100);

hist(df_test.last, 100);
