==> JuliaAcademy/DataFrames/6_next_steps.ipynb <==
# Final examples
Joining and reshaping data

using DataFrames, CSV, Pipe, Unitful, Dates

# Load rainfall data
rainfall_long = CSV.File("rainfall_forecast.csv") |> DataFrame

transform!(rainfall_long, :rainfall => x -> x .* u"mm", renamecols=false)

eltype.(eachcol(rainfall_long))

rainfall_wide = unstack(rainfall_long, :date, :city, :rainfall)

# No problem with special character
rainfall_wide."Ełk"

# Fill missing data and missing days with 0 mm
# 1. Create array of all days
all_days = DataFrame( "date" => Date.(2020,11,16:22))

rainfall_all = @pipe leftjoin(all_days, rainfall_wide, on=:date) |> coalesce.(_, 0.0u"mm")

sort(rainfall_all, :date)
